requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.9.0
eventlet==0.33.3
requests==2.31.0
pandas==2.0.3
matplotlib==3.7.2
numpy==1.24.3
uuid==1.30
python-dotenv==1.0.0
flask-cors==4.0.0
